// src/modules/offer/types/mail.status.type.ts
export enum MailStatusType {
  CUSTOMER_NEW_OFFER_REQUEST = 'CUSTOMER_NEW_OFFER_REQUEST',
  CUSTOMER_REQUEST_CORRECTION = 'CUSTOMER_REQUEST_CORRECTION',
  CUSTOMER_REQUEST_CORRECTION_REMINDED = 'CUSTOMER_REQUEST_CORRECTION_REMINDED',
  CUSTOMER_REQUEST_CORRECTION_FROM_CUSTOMER = 'CUSTOMER_REQUEST_CORRECTION_FROM_CUSTOMER',
  FEE_REQUEST_FOR_OFFER = 'FEE_REQUEST_FOR_OFFER',
  FEE_REQUEST_FOR_OFFER_REMINDED = 'FEE_REQUEST_FOR_OFFER_REMINDED',
  SUPPLIER_NEW_OFFER = 'SUPPLIER_NEW_OFFER',
  CUSTOMER_OFFER_RESULT = 'CUSTOMER_OFFER_RESULT',
  WAITING_FEE_FOR_OFFER = 'WAITING_FEE_FOR_OFFER',
  INTERNAL_INFORMATION = 'INTERNAL_INFORMATION',
  FILE_NOTIFICATION = 'FILE_NOTIFICATION',
  OTHER = 'OTHER',
}